Updates the CMakeLists.txt file to work on a Linux distribution.     Could still be improved.
Attempt fix a warning because of a pointer to an array used instead of the first cell of said array.

Specifies "static" to several consts so there is no misunderstanidng as the permanent location status of the objects.

Fixes a potential buffer overflow.

Fixes a pointless and potentially confusing "else" (removed it)

Fixes an array size mismatching the size of the expected buffer in a called function (512 vs 48KB)

Fixes an incorrectly defined struct (missing a ; at the end)